server/scrapers/requirements.txt
pytz==2019.3
requests==2.22.0
pymysql==0.10.1
python_dateutil==2.8.1
PyYAML==5.3.1
sodapy==2.1.0
